+++ src/DotNext.Tests/Threading/AsyncCounterTests.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
-using System.Threading;
 using System.Threading.Tasks;
 using Xunit;