+++ src/DotNext.Tests/Threading/Tasks/SynchronizationTests.cs
@@ -4,7 +4,6 @@ using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
-using static System.Threading.Timeout;
 
 namespace DotNext.Threading.Tasks
 {