+++ src/DotNext.Unsafe/Buffers/IUnmanagedMemoryOwner.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Buffers;
 
 namespace DotNext.Buffers
 {
@@ -11,14 +10,14 @@ namespace DotNext.Buffers
     /// unmanaged memory.
     /// </summary>
     /// <typeparam name="T">The type of elements to store in memory.</typeparam>
-    public interface IUnmanagedMemoryOwner<T> : IMemoryOwner<T>, IUnmanagedArray<T>
+    public interface IUnmanagedMemoryOwner<T> : IUnmanagedMemory<T>, IUnmanagedArray<T>
         where T : unmanaged
     {
         /// <summary>
         /// Resizes a block of memory represented by this instance.
         /// </summary>
         /// <remarks>
-        /// This method is dangerous becase it invalidates all buffers returned by <see cref="IMemoryOwner{T}.Memory"/> property.
+        /// This method is dangerous becase it invalidates all buffers returned by <see cref="System.Buffers.IMemoryOwner{T}.Memory"/> property.
         /// </remarks>
         /// <param name="length">The new number of elements in the unmanaged array.</param>
         /// <exception cref="ObjectDisposedException">The underlying unmanaged memory is released.</exception>